backend/pyproject.toml
[project]
name = "pelicanone-backend"
version = "0.1.0"
description = "PelicanOne 2.0 backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "redis>=5.0",
    "rq>=1.15",
    "httpx>=0.27",
    "pyjwt>=2.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "aiosqlite>=0.20",
    "ruff>=0.3",
    "black>=24.2",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
